Added segmented controllers to storybook
